utils/toolbox: adjust equilateral triangle coordinates
The triangle now has its center of gravity on (0,0) and the size hint sz
is replaced with the diameter of the circle in which the triangle fits.
This is more meaningful than the altitude (sz is h=3r/2, or 3d/4). Note
that the previous triangle was still equilateral (and thus correct), but
harder to work with.

Some tests make use of these coordinates so they are updated. The
references still make sense (and are sometimes better) with this
triangle.