pass: reset all unused samplers before executing the pipeline
Fixes graphics errors happening when trying to bind an image with a
layout different from the previously bound image. These graphics errors
are due to the the previous image planes still being bound while their
corresponding textures have been de-allocated by the hwupload module.

This scenario happens if a media changes its output format mid stream
due a change in its bitstream (provoking a codec re-initialization).

The hwconv module which currently relies on ngl_texvideo() is left as is
as it only supports input images with the same layout as the one
provided during its initialization. If a change of format occurs, the
whole module is re-initialized by the hwupload module.