internal: add distance map API
This is an interface similar to the internal atlas API, except it does
take a path (lines and curves) instead of a bitmap as argument, and
output a signed distance field.

This API is meant to be used for path and text rendering in future
commits. It will unlock all kind of fast advanced effects, as well as
smooth rendering at high resolution.

The rasterization of the path into that SDF is done in a dedicated
fragment shader. The shader accepts a bunch of bézier cubic points as
input, and perform a 5th degree polynomial resolution (because we're
solving a distance squared of these bézier cubic curves) using
Aberth-Ehrlich root finding algorithm.

To avoid all kind of nastiness with regards to the winding number (the
algorithm used to defined whether we are inside or outside the shape)
because of float arithmetic, we use a novel approach inspired by Eric
Lengyel for quadratic bézier, except we extend it to cubics using
various tricks. This will be documented in more details in the future.

Changes from Matthieu:
- simplification of the RT layout
- caps probing for the output texture format
- switch to the dynamic uniform buffers the draw calls
- help fixing a few glitches caused by bugs in the distmap fragment

Signed-off-by: Matthieu Bouron <matthieu.bouron@gmail.com>