hwupload_videotoolbox_ios_gl: fix and simplify the direct rendering detection
This commits fixes the direct rendering detection for RGBA/BGRA
Videotoolbox buffers by scoping the supported image layouts check to the
NV12 buffers codepath and generalizing the mipmap filter check to all
kind of buffers.

This commit also fixes a wording issue regarding the type of handle we
are dealing with (IOSurface (macOS) -> Videotoolbox textures (iOS)).