hwupload: change hwmap class selection logic
The current hwmap class selection logic is a based on a direct mapping
(per backend) between the hwmap classes and all the sxplayer pixel
formats.

The new logic, introduced by this commit, is based on an array of hwmap
classes that only supports hardware-accelerated pixel formats and a
fallback on the common hwmap class if the frame pixel format is not
supported by any of the hwmap classes declared in the array.

This new logic is more concise and will ease the addition of new
backends.