hmap: make hmap browsing insert-ordered
When pgcraft build the shader string using its resources, the order in
the declared slots (inputs, outputs, uniforms, ...) is "random"
(essentially, it's based on the CRC32 hashing function from hmap). This
makes debugging and testing clumsy.

A side effect of this commit is that ngli_hmap_next() is much simpler.