geometry,pass: remove creation of internal nodes in geometries
Pass is working with a geometry_priv structure, shared amongst all
geometries: Geometry (where all attributes are configurable through user
nodes) and the Triangle, Quad, Circle bare geometries where the
attributes are internal buffers.

Before this commit, the approach is to make the bare geometries act like
the generic Geometry (by creating internal nodes through the high level
public API) so that pass works with a common interface for all
geometries. This creates a fragile and awkward cross-levels use of the
API (typically the attach/detach ctx logic) which prevents refactoring.

The new approach in this commit is to invert the model: the bare
geometries just work with simple GPU buffers, while the node Geometry is now
managing its buffer nodes (instead of the pass module).

This causes the pass module to work with GPU buffers and their layout
(still read from the geometry private context) instead of the
corresponding nodes (which don't exist anymore in the case of the bare
geometries).

The changes in the bare geometries is pretty straightforward: it simply
uses the recently introduced internal geometry API to build the GPU
buffers, and the node management goes away.

For the Geometry node, changes are a bit more complex because it is
now managing its own child nodes. Similar to what's found in pass for
the other attributes, there is a special code path for buffer nodes that
reference a block, and there is the ref/update/upload/unref mechanisms.

In pass, the indices code gets simplified (since it's now entirely
managed by geometries), the geometry update is now called, and the
vertices/uvcoords/normals attributes are treated as special cases
(because they're not nodes anymore like other attributes, but buffer +
layout couples).