utils/misc: add a rng into the scene config
This RNG is a helper for getting random values within scene building
functions. Using the global random state is not a good idea because
concurrent calls to a function (which typically happens with the hooks)
will lead to a non-deterministic outcome as the same common global
random state is going to be consumed from concurrent function calls.