params: allow setting a node on some elementary param types
We make it possible for some parameters to accept a node instead of a
direct value. This allow providing a less awkward interface for some
node that can work either with a direct value or an animated one:
typically transforms.

When the flag is set on a parameter, the corresponding field private
structure must be a struct ngl_node *, followed by the "raw" field.
Doing it the other way around is problematic due to the alignment rules.
With the pointer first, we can assume the alignment.