media,texture: fix shared media guard
The logic introduced by c4f812a5a03823992430bff3fe77c7ef0488c39d was
incomplete as it assumed that ngli_node_prepare() was called only once
per parent node. This is not the case as ngli_node_prepare() is called
once per parent per rendering branch. The code was also not triggered
from proper rendering graphs as the Render and RT nodes (which are the
only possibles parent of a Texture node) were not properly forwarding
the prepare callback to its children (this will be addressed in the next
commits).

Since this guard cannot be implemented in the prepare callback for the
reasons mentioned above, this commit moves it to the Texture2D init
callback.