params: rename param names exposed to specs and doc
We are progressively moving away from abstract types ("int", "unsigned",
"float", "double") to explicitly sized types such as "i32", "u32",
"f32", "f64". In addition to aesthetic, it also aligns with assumptions
currently made in the code about sizes and how these types are actually
tested in practice.

We take the opportunity of this rename to switch to snake_case the list
and dict based parameters, for consistency.

This work prepares the incoming commits with a new API following this
scheme. That way, we will have an API consistent with the specs and the
doc.