pgcraft: simplify buffer binding logic
The current logic, before this commit, explicitly disables explicit
bindings in the shader even if they are supported by the context. This
is due to the way the uniform buffers are handled when probing the
resources of the resulting OpenGL program; the uniform buffer probing
routine is setting the uniform buffers bindings to arbitrary values
forcing pgcraft to use the shader resource description to properly set
the pipeline_buffer_desc.binding field.

This commit changes the current logic by letting pgcraft always decide
the buffers bindings even if the context does not support explicit
bindings. On such context, we execute a post processing pass on the
program to enforce the uniform buffer bindings according to the pgcraft
resource description.