pgcraft: simplify attribute binding logic
Simplifies the attribute logic binding by letting pgcraft always decide
the attribute locations even if the context does not support explicit
bindings. On such context, we execute a post processing pass on the
program to enforce the attribute locations according to the pgcraft
resource description.

According to the glBindAttribLocation documentation, changing the vertex
attribute locations requires the program to be linked again:

> Attribute bindings do not go into effect until glLinkProgram is called.