texture: add Texture2D.clamp_video parameter
Sampling is not guaranteed to be normalized. This at least happens
sometimes with some videos with VAAPI and Android OES sampling.

While harmless most of the time, the overflow can have unexpected
effects for the users.

It is not enabled by default because it doesn't happen that often, and
the overflow may actually be considered a feature in some cases.