nodes: release nodes in the right order (bottom-up)
Prefetching the graph resources for a given time must prefetch the nodes
in the bottom-up order where the leaf nodes (textures, buffers, ...) are
prefetched first, then their parents up to the root node.

On the other hand, releasing the graph resources for a given time must
release the nodes in the top-bottom order where the leaf nodes must be
released last.

                           [A]
                            |
                       .____.____.
                       |    |    |
                      [B]  [C]  [D]
                       |   /     |
                       |  /      |
                       | /      [F]
                       [E]

Without this commit the above graph would have released the nodes in the
following order: E, B, C, F, D, A which is incorrect.

This commit simply reverses the previous order so it becomes A, D, F, C,
B, E which is correct as the leaf nodes are now released last.