utils/resourcetracker: blacklist config file
Currently, the configuration is potentially re-written every second (see
_config_timer in config.py). "potentially" includes notably at every
scene change event, which actually happens every time the user interacts
with a custom scene option widget in the GUI (because pkg/module/scene
settings are re-stored in the config every time). This means that if the
1-second GUI config timeout overlaps concurrently with the scene
construction (because com.py:query_inplace is called from multiple
QThread used by the HooksController), the config file will be tracked
and included in the file list to monitor. In such event, a reload event
will follow immediately, causing all the user custom options to be
reset.

In practice this happens quite frequently when playing with custom scene
options, in particular with widgets such as sliders which spams scene
change events.