hmap: fix over-fixing references on delete
Following is a scenario where the issue could happen. Let's say we have
the following entries in a given bucket (first line is the index in the
allocated entries array):

       0    1    2      3       4
    | foo | X | bar | lorem | ipsum |

Let's assume `X` is the removed element, which means every entry (in the
whole hmap) which references `bar`, `lorem` or `ipsum` through their
prev or next reference is off by one. Indeed, after the memmove, the new
bucket.entries is layed out like this in memory:

       0     1      2       3
    | foo | bar | lorem | ipsum |

In fix_refs(), we loop over `bar`, `lorem` and `ipsum` (in this order),
and for each of these entry we use fixed_entry_from_ref() on their prev
and next references to obtain the entries that point to them. When we
have these entries, we can fix their complementary next or prev
references so that they point on the correct current index.

Let's assume now that bar.prev links to `lorem` (it could be bar.next,
it makes no difference). The reference that creates this link has a
value of 3, which is invalid in the current state (it would point to
`ipsum`), so we need to adjust by -1. This is controlled by the
condition "ref.bucket_id == removed.bucket_id && ref.entry_id >
removed.entry_id".

So at the first iteration, we point on `lorem` correctly thanks to the
-1 live fix. Now we make one change: the complementary reference
lorem.next (which has a value of 2 instead of 1) is decremented.

The fix_refs() loop continues and we arrive at the `lorem` entry. The
lorem.next reference has the value 1 (because it was just fixed). Before
this commit, fixed_entry_from_ref(next) would decrement this value
again, and the reference would be broken: in this case pointing on `foo`
instead of `bar`. In other cases we could have a negative value (for
example if the entry #0 instead of #1 was removed) causing a crash.

The added condition in this commit prevents the decrement from happening
twice. We know for sure that if the current entry points to a back
reference, it means a previous iteration in fix_refs() passing through
that back reference made sure the current one was already fixed, so we
must skip the offsetting.

In practice, the bug doesn't happen often for two reasons. First, the
hash map is so large in practice that multiple entries in the same
bucket is not that common (it happens more in the test where the hmap
size is very small). Secondly, deletes don't actually happen very often
(almost never) as scene construction is mainly a growing operation. The
only common deleting operation happening is the hmap freep, but it
doesn't go through this code path, it just destroys everything directly.