rtt: set the depth attachment store operation depending on the number of render pass interruption
Before this commit, when we knew that the render pass was interrupted we
were setting the depth attachment store operation to
NGLI_STORE_OP_DONTCARE for depth attachments that were not user supplied
textures. This is not valid for the following use case (when there are
more than one interruption):

                           [RTT]
                             |
             .-------.-------.-------.-------.
             |       |       |       |       |
             |       |       |       |       |
             |       |       |       |       |
          [Render] [RTT]  [Render] [RTT]  [Render]
                                  ^
                                  |
                                  |
                 The depth attachment is discarded here,
                 which is invalid because the next
                 render can potentially use it

After this commit, we set the depth attachment store operation
depending on the number of render pass interruptions. We can only
discard the depth attachment if there is only one interruption. See the
following scenarios:

RTT with no render pass interruption (we use only one rendertarget with
LOAD_OP_CLEAR for all attachments and STORE_OP_STORE for the color
attachments and STORE_OP_DONTCARE for the depth attachment if it is
not a user supplied texture):

                           [RTT]
                             |
                     .-------.
                     |       |
                     |       |
                     |       |
                  [Render] [RTT]
                          ^
                          |
                          |
                 We can discard the
                 depth attachment here

RTT with one render pass interruption (we use two rendertargets, the
first one with LOAD_OP_CLEAR for all attachments and STORE_OP_STORE for
all attachments, the second one with LOAD_OP_LOAD for all attachments
and STORE_OP_STORE for all color attachments and STORE_OP_DONTCARE for
the depth attachment if it is not a user supplied texture):

                           [RTT]
                             |
                     .-------.-------.
                     |       |       |
                     |       |       |
                     |       |       |
                  [Render] [RTT]  [Render]
                                          ^
                                          |
                                          |
                                 We can discard the
                                 depth attachment here

RTT with two render pass interruptions (we use two rendertargets, the
first one with LOAD_OP_CLEAR for all attachments and STORE_OP_STORE for
all attachments, the second one with LOAD_OP_LOAD for all attachments
and STORE_OP_STORE for all attachments):

                           [RTT]
                             |
             .-------.-------.-------.-------.
             |       |       |       |       |
             |       |       |       |       |
             |       |       |       |       |
          [Render] [RTT]  [Render] [RTT]  [Render]
                                                 ^
                                                 |
                                                 |
                                        We can discard the depth
                                        attachment here but we
                                        currently don't do it because
                                        that would require to generate
                                        a third rendertarget with
                                        another set of load/store
                                        operations [1]

[1] We don't want to optimize such scenario as interrupting render
passes implies performance penalties.