rendertarget_gl: clear depth attachment of wrapped rendertargets
This commit makes sure the depth attachment of the rendertarget wrapping
the default OpenGL framebuffer is cleared at the beginning of the render
pass if requested. This fixes onscreen rendering when depth testing is
enabled.

Fixes a regression introduced by commit
91348e51263a7f9820fdcc0fdc4acf35f8aed4a8 (which removed explicit clear
commands in gl_begin_draw()).

The regression was not caught by the tests because they are ran
offscreen.