api: dispatch the reset command to the rendering thread in configure_ios()
The graphics resources associated with the graph are owned by the OpenGL
context that is current within the rendering thread. Dispatching the
reset command to the rendering thread ensure that the OpenGL calls
responsible to freeing up the resources are made from the proper thread.

Executing the reset command (and its associated OpenGL calls) on the
wrong thread can lead to several issues, including deleting wrong
resources tied to another OpenGL context (if bound to the current
thread) or crashes if no OpenGL context is bound to the current thread.
This was somehow not an issue with ngl-desktop on macOS. From what I
observed on this platform, once an OpenGL context has been initialized
on the current thread, calling OpenGL functions won't crash anymore even
though no OpenGL context is current.