internal: add eval API
This API will become more and more useful as we get smarter nodes,
typically nodes that implement a custom shader behind the scenes. Some
expressions within these shaders will be out of reach to the user, so
this eval offers a way to compute CPU wise some expressions that will be
fed to these "black boxes".

The `^` operator is currently not implemented; it is usually present as
a power operator, but if the eval API evolves to support integers, it
means we won't have that character available for a xor operator.

It is also slightly more complex to implement because contrary to other
operators it is right-associated instead of left-associated (a^b^c
is equivalent to a^(b^c), not (a^b)^c). As a result, we simply expose a
pow() function.

The code could potentially evolve in several directions. Aside from the
support for integer based types, we could also support vectorial data,
both as input and output. We could also explore the idea of
introspecting into data buffers or more complex structures. Another
improvement could be to try to use a just-in-time compiler for the
evaluation (`ngli_eval_run` + `eval_operator_pop` basically), such as
libjit or eventually llvm.

One last note about the #pragma function: like the comment says, this is
only for MSVC to compile. We are currently listing all the functions,
even for those without intrinsics in the current version (because we
never know if they're not going to be implemented in a future version,
or weren't available in a older one), which means the compilation prints
out some warnings.

An alternative would be to reimplement a wrapper for all these
functions, but then this would cause a double function jump in many
cases.