tests/media: add timeranges_rtt test
This test exercises mainly the prefetch/release mechanism which is used
by RTTs and Medias. The test is located in tests/media.py because it's
the only set of tests that has access to the media. The scene has a
diamond shape where the 2 branches turns on and off, with a phase of
overlap.