params: adjust string truncation heuristic
This code is only for the dot representation. Using the presence of '/'
to determine whether the string is a path is not a good enough heuristic
anymore with the introduction of the Eval nodes. Indeed, a simple
expression such as "sin(t)/2" will lead to displaying a confusing "2" in
the dot graph.

The new heuristic just truncates the beginning of the string and keep
the last 32 characters. We keep the end so that it stays useful for
representing long paths.