glstate: replace ngli_glstate_probe() with ngli_glstate_reset()
This function will be useful to reset the OpenGL state when wrapping
external OpenGL contexts as we want to avoid probing the OpenGL state
before each call to ngl_draw().

This commit also replaces ngli_glstate_probe() with this new function so
that instead of probing the OpenGL state after context creation to
initialize the glstate structure we now reset the OpenGL state to its
default and initialize the glstate structure accordingly.