glstate: inline init_state() and honor_state() in ngli_glstate_update()
This commit inlines init_state() and honor_state() in
ngli_glstate_update(). It also removes the use of memcmp() to detect
graphics state changes as it was not working properly. As a matter of
fact, the current state was compared to the new desired state based on
an incomplete glstate structure lacking the scissor, viewport and
program_id fields. This incomplete glstate structure was then copied to
the current glstate structure discarding the previous state of
the scissor, viewport and program_id.