api: add live controls API
The goal of this API is to allow scenes to expose specific live controls
for the end user to play with. While it was already possible for scene
building functions to return an additional list of uniforms, in this API
the controls survive the serialisation process (the user doesn't need to
tinker with custom transmission layers).

In practice, this means `ngl-desktop` (and `ngl-python`) can use this
API to grab the scene controls, and expose an UI, typically using an
immediate mode GUI such as DearImgui or Nuklear.

Various alternate design were considered during development. One of them
was to have the API work with an instantiated context. In this attempt,
the uniform nodes were registering themselves into a context live
controls hash map, and the ngl_livectls_get() API was simply reading
this hash map. It was also possible to go through the flat node lists
built during the init process, which made the code more straightforward.
While the code was simpler on node.gl side, it was causing an
architectural nightmare for the users: the usual design of a graphical
app using this API is to have an UI rendering widget that communicates
down information to its underlying renderer asynchronously, but that
widget is rarely aware of the renderer internals. So having to
interrogate the renderer in order to display UI elements after it has
been configured implied a double asynchronous communication. In Qt/QML,
this is typically extremely challenging to build, revealing the design
shortcoming of such an approach.