pynodegl: expose live controls API
The way we transmit the livectl_info from Python to Cython is a bit
clumsy (the Python side fills up a dictionary in the Cython module after
importing it), but I couldn't find a better way. The issue is that the
Cython side needs to instantiate using node classes defined dynamically
on the Python side.

Note that `ngl_livectl.node_type` is not exposed in the returned
dictionary because it is hardly usable: the `NGL_NODE_*` defines are
not exposed to the Python (one needs to access the private
`pynodegl._ngl` module). Instead, we use the node class name.