nodes: add UserSelect
While UserSwitch is useful to turn on and off some sort of overlay on
the scene, it is clumsy to use when deciding to pick one of multiple
branches because it implies simultaneous operations (one enable, one
disable). When exposed through live controls, it's even more problematic
if these branches are supposed to be exclusive.