text: split alpha/opacity component out of colors
The alpha component of a RGBA color is ambiguous because we don't know
if the color is supposed to be pre-multiplied or not. With a separate
opacity settings, we can choose how to interpret it. Currently, we
consider it to be equivalent to straight alpha, but we are going to move
to the proper opacity semantic and premultiply it internally.

Separating the opacity from the color also has the benefit for the user
to more easily build different animations.