text: configure the blending of the background as well
The background color is premultiplied and meant to be blend as such.

The tests are updated because now the alpha channel of the default FBO
is not messed up anymore.

It was considered leaving the blending mode up to the user (similarly to
how it's done in Render* nodes with a user parameter), but a wide
variety of blendings wouldn't work as expected because the Text node has
the particularity of drawing 2 layers (the background, then the
characters).

We could have removed the ability to draw the background layer in the
node and leave that up to the user. It would have worked in the current
state because the text characters are drawn in a single quad. But sooner
or later, we are going to have text drawn using a bunch of separate
quads. And even if they are going to be drawn in the same render pass,
it won't be a single big area usable for many blending modes: all the
empty spaces not drawn will be messed up. Not even mentioning the case
of overlapping quads.

This means that if users want to use the Text as a mask, they have to
use an intermediate RTT and use a RenderTexture with the appropriate
blending mode (just like currently).