rendertarget_gl: unconditionally handle the depth/stencil load/store operations
We can unconditionally set the depth/stencil clear flags and invalidate
attachments when wrapping a framebuffer as glClear() and
glInvalidateFramebuffer() ignore attachments that do not exist in the
bound framebuffer.

This fixes the clearing (if requested) of the swapchain depth/stencil
attachment on context not supporting the glClearBufferf{i,v} functions
(OpenGLES 2.0). The issue can be reproduced with:

export MESA_GLES_VERSION_OVERRIDE=2.0
ngl-python -b opengles -u tests/depth_stencil.py depth_stencil_depth

This also fixes the invalidation (if requested) of the swapchain
depth/stencil attachment. This cannot be reproduced as we never set the
depth/stencil attachment store operation to DONT_CARE when wrapping a
framebuffer.