build: add check kwarg to run_command() call
Sets the check argument to true as the related commands are not meant to
fail.

Fixes the following warning:

WARNING: You should add the boolean check kwarg to the run_command call.
It currently defaults to false, but it will default to true in future
releases of meson.
See also: https://github.com/mesonbuild/meson/issues/9300