rendertarget: remove ngli_rendertarget_read_pixels()
Currently ngli_rendertarget_read_pixels() is only used in the CPU
capture callback of gpu_ctx_gl and can be inlined there. The new logic
is also simpler compared to the previous one since the code cannot be
called in the middle of a render pass anymore.

Moreover, ngli_rendertarget_read_pixels() is not great API-wise as it
only supports reading the content of the first color attachment and is
too tied to OpenGL (which in most cases requires a framebuffer to read
back the pixels of a texture whereas modern backends can directly work
with textures). A similar functionality should be implemented as
ngli_texture_download() or something similar if one needs to download
the data of a texture.