buffer: transform buffer_priv into buffer_info
Mainly this commit is about buffer_priv being renamed to buffer_info.
But to prevent an intermediate commit that rename everything in
node_buffer.c and then back again, the buffer_priv private struct is
introduced at the same time in that file.

The fp parameter is the only thing exclusive to the block node, so it is
moved into that buffer_priv structure.

After this commit, the 3 types of buffer nodes (Buffer*, AnimatedBuffer*
and StreamedBuffer*) all have their own dedicated private context and
private options. All their private context starts with a buffer_info
struct (with associated compile check) so that external units can
interpret their private context as the generic buffer_info struct.
Everything else is opaque externally.