feature: add NGLI_FEATURE_TEXTURE_{FLOAT,HALF_FLOAT}_RENDERABLE
Theses features indicate that float/half-float textures can be used as
color attachments to a render pass. The suffix _RENDERABLE is used
instead of COLOR_BUFFER as COLOR_BUFFER is ambiguous and tied to the
corresponding OpenGL extensions. The naming *-renderable is also used in
the OpenGL specification to qualify this particular feature.