pgcraft: use hwmap to inspect available image layouts
The available image layouts depend on the backend and platform, and so
we don't need (nor want) to inject code that tries to sample every
combination. Instead of having to keep pgcraft synchronized with how
hwmaps are registered, we make pgcraft interrogate what's currently
available. This makes the code more robust to changes happening in
backends and hwmaps.