pass: explicitly update all uniforms
Before this commit, pass.c was relying on a pipeline_gl behavior which
updates automatically all the uniforms for which a data pointer has been
associated with using ngli_pipeline_set_resources(). This behavior is
specific to pipeline_gl and won't be supported on modern backend
implementations (which don't support uniforms).

This commit makes pass.c construct and use an explicit list of uniforms
and their data. This makes pass.c compatible with the upcoming changes
that add support for mapping single uniforms to dedicated uniform buffers
on modern backends.

This commit also makes the code consistent with node_renderother.c.