nodes: add RenderPath node
The path shading code is kept in a separate path.glsl helper for use in
other modules.

Currently the most obvious limitation is that the geometry cannot be
specified (it's a full screen quad). Adding the ability to specify a
geometry should be doable but will require careful thinking with regards
to the scaling anchor. Allowing only quads at first might be a solution.