utils: flag specific code blocks with custom formatting
This "fmt: on/off" tagging is used by Black to enable/disable
formatting. We use them in places were vertical indent or
linebreaks vs packing makes sense (matrices or pairing of arguments).

For the blocks with single quotes, we are changing them to double ones
because Black will not change them anymore.