utils/qml: add QML helpers
The NodeGLWidget is dynamically injected into the Player so that the
player can be tested with the QML scene viewer: `qml6 Player.qml`
without requiring a complete app/backend.

The whole player logic (timing etc) is embedded into Player.qml so
taking the NodeGLWidget out of the equation, it is a perfectly re-usable
component.

On the backend side, the NodeGLWidget basically derivates a framebuffer,
backed by a custom renderer (see ngl_widget.py). Since they are not
necessarily running in the same thread, the commmunication between the
two is managed through the synchronize() mechanism. The widget is
expected to be connected to the time and scene change events, which are
passed down (along with the live changes events) to the renderer which
honors them.

We make sure the live change events are executed in the renderer thread
so that they are properly serialized between draw calls.

For now, all this mechanic only works with the OpenGL backend. A Vulkan
version is likely to require a complete different model.