utils: add ngl-diff tool
This tool allows analyzing differences between medias. It's the first
tool we have making use of the external OpenGL support and our QML
helpers.

It is basic but functional. Zooming (mousewheel) and panning (right
click) are supported, as well as moving the split position (left click)
and orientation (vertical/horizontal).

The UI can filter RGBA channels to study differences on specific ones,
or exclude some. For the alpha channel, there is the ability to
interpret it as premultiplied or not.

The Diff mode uses an exponential that boosts the small changes (from
blue to red). Black means no change. The associated slider controls the
threshold for considering differences (from 0 to 10%).

If medias have different properties, they will be highlighted in red the
UI.