project: add pyproject.toml for black and isort
That way, running black and isort use the appropriate configuration by
default.

In addition to the line length settings, The black profile for isort
allow us to run black and isort in any order.

While not recommended, isort configuration could be added to the
setup.cfg, but black doesn't support it so we take the latest trending
bikeshed and use pyproject.toml.