math_utils: rework a bunch of vec4 based operations
The rationale behind this change is not cosmetic but motivated by the
removal of the compiler aliasing constraint. In all these functions, the
destination can alias with the source(s) because the functions support
in-place transformations (by passing identical pointers).

By being explicit about it using a local copy, we reveal to the compiler
that these data must be treated separately. Doing this becomes relevant
starting data length of 4x32-bit floats because the compiler can start
making use of SIMD.

If we were to study the change on ngli_vec4_add() on x86 with GCC
(starting -O3 optimisation level, which we used in release build), the
assembly goes from:

    movss   xmm0, DWORD PTR [rsi]
    addss   xmm0, DWORD PTR [rdx]
    movss   DWORD PTR [rdi], xmm0
    movss   xmm0, DWORD PTR [rsi+4]
    addss   xmm0, DWORD PTR [rdx+4]
    movss   DWORD PTR [rdi+4], xmm0
    movss   xmm0, DWORD PTR [rsi+8]
    addss   xmm0, DWORD PTR [rdx+8]
    movss   DWORD PTR [rdi+8], xmm0
    movss   xmm0, DWORD PTR [rsi+12]
    addss   xmm0, DWORD PTR [rdx+12]
    movss   DWORD PTR [rdi+12], xmm0
    ret

...To:

    movups  xmm0, XMMWORD PTR [rsi]
    movups  xmm1, XMMWORD PTR [rdx]
    addps   xmm0, xmm1
    movups  XMMWORD PTR [rdi], xmm0
    ret

A similar change can be observed with Clang (starting -O2) optimisation
level.

An alternative to this approach resulting in the same outcome would be
to use the `restrict` C99 keyword, but we would actually be violating
the contract (because we do actually pass down aliased pointers),
leading to undefined behaviour.

Note that these changes are irrelevant in ngli_mat4_mul() and
ngli_mat4_mul_vec4() because we are already using an intermediate
destination buffer. The important change here is the separation between
the reading and the writing operations on their aliased data.